create subclasses and specs for card_type and valid checks
