+++ app/card_number_processor.rb
@@ -5,24 +5,11 @@ class CardNumberProcessor
   end
   
   def card_type
-    return 'Unknown' unless @card_number.start_with?('6014')
-
-    length = @card_number.length
-
-    case valid_card_number_format?(@card_number)
-    when black_card?(length)
-      return 'Flybuys Black'
-    when red_card?(length)
-      return 'Flybuys Red'
-    when green_card?(length)
-      return 'Flybuys Green'
-    else
-      return 'Flybuys Blue'
-    end
+    CardTypeChecker.new(@card_number).determine_card_type
   end
   
   def valid?
-    return validate_card_number(@card_number)
+    CardValidityChecker.new(@card_number).check_validity
   end
   
   private
@@ -34,40 +21,6 @@ class CardNumberProcessor
 
   def valid_card_number_format?(card_number)
     card_number.match?(/\A\d{16,17}\z/)
-  end
-
-  def black_card?(length)
-    (@card_number[4] == '1') && (length == 16 || length == 17)
-  end
-  
-  def red_card?(length)
-    @card_number.start_with?('6014352') && length == 16
-  end
-  
-  def green_card?(length)
-    @card_number.start_with?('601435552') &&
-      (6..9).include?(@card_number[9].to_i) &&
-      length == 16
-  end
-
-  def validate_card_number(card_number)
-    digits = card_number.chars.map(&:to_i)
-    doubled_digits = []
-  
-    digits.reverse.each_with_index do |digit, index|
-      if index.odd?
-        doubled = digit * 2
-        doubled = doubled.digits.sum if doubled > 9
-        doubled_digits << doubled
-      else
-        doubled_digits << digit
-      end
-    end
-  
-    total = doubled_digits.sum
-    valid = (total % 10).zero? ? 'valid' : 'invalid'
-
-    return valid
   end
 end
