+++ spec/card_number_processor_spec.rb
@@ -1,14 +1,12 @@
 require_relative '../app/card_number_processor'
+require_relative '../app/card_type_checker'
+require_relative '../app/card_validity_checker'
 
 RSpec.describe CardNumberProcessor do
-  let(:valid_black_card_number_16_digits) { '6014111100033006' }
-  let(:valid_black_card_number_17_digits) { '60141016700078611' }
+  let(:valid_black_card_number) { '6014111100033006' }
   let(:invalid_black_card_number) { '6014152705006141' }
-  let(:valid_red_card_number) { '6014352700000140' }
-  let(:valid_green_card_number) { '6014355526000020' }
-  let(:valid_blue_card_number) { '6014709045001234' }
   let(:unknown_card_number) { '6013111111111111' }
-  
+
   describe '#initialize' do
     let(:invalid_input_type) {12345}
     let(:invalid_input_length) {'12345'}
@@ -16,7 +14,7 @@ RSpec.describe CardNumberProcessor do
 
     context 'with a valid card number' do
       it 'initializes a CardNumberProcessor object' do
-        expect { CardNumberProcessor.new(valid_red_card_number) }.not_to raise_error
+        expect { CardNumberProcessor.new(valid_black_card_number) }.not_to raise_error
       end
     end
 
@@ -36,58 +34,26 @@ RSpec.describe CardNumberProcessor do
   end
 
   describe '#card_type' do
-    context 'when a valid card number' do
-      context 'of Flybuys Black type' do
-        it 'returns the correct type for a 16 digit number' do
-          processor = CardNumberProcessor.new(valid_black_card_number_16_digits)
-
-          expect(processor.card_type).to eq('Flybuys Black')
-        end
-      
-        it 'returns the correct type for a 17 digit number' do
-          processor = CardNumberProcessor.new(valid_black_card_number_17_digits)
-
-          expect(processor.card_type).to eq('Flybuys Black')
-        end
-      end
-
-      it 'returns the correct type for Flybuys Red numbers' do
-        processor = CardNumberProcessor.new(valid_red_card_number)
-
-        expect(processor.card_type).to eq('Flybuys Red')
-      end
-
-      it 'returns the correct type for Flybuys Green numbers' do
-        processor = CardNumberProcessor.new(valid_green_card_number)
-
-        expect(processor.card_type).to eq('Flybuys Green')
-      end
-
-      it 'returns the correct type for Flybuys Blue numbers' do
-        processor = CardNumberProcessor.new(valid_blue_card_number)
-
-        expect(processor.card_type).to eq('Flybuys Blue')
-      end
+    it 'returns the correct card type of a known card number' do
+      card_type_checker = CardTypeChecker.new(valid_black_card_number)
+      expect(card_type_checker.determine_card_type).to eq('Flybuys Black')
     end
 
-    it 'handles unknown card numbers appropriately' do
-      processor = CardNumberProcessor.new(unknown_card_number)
-
-      expect(processor.card_type).to eq('Unknown')
+    it 'returns the correct card type of an unknown card number' do
+      card_type_checker = CardTypeChecker.new(unknown_card_number)
+      expect(card_type_checker.determine_card_type).to eq('Unknown')
     end
   end
 
   describe '#valid?' do
-    it 'returns "valid" for a valid card number' do
-      processor = CardNumberProcessor.new(valid_black_card_number_16_digits)
-
-      expect(processor.valid?).to eq('valid')
+    it 'returns valid for a valid card number' do
+      card_validity_checker = CardValidityChecker.new(valid_black_card_number)
+      expect(card_validity_checker.check_validity).to eq('valid')
     end
 
-    it 'handles invalid card numbers appropriately' do
-      processor = CardNumberProcessor.new(invalid_black_card_number)
-
-      expect(processor.valid?).to eq('invalid')
+    it 'returns invalid for an ivalid card number' do
+      card_validity_checker = CardValidityChecker.new(invalid_black_card_number)
+      expect(card_validity_checker.check_validity).to eq('invalid')
     end
   end
 end
